Fundaments - Dungeon Survivors
This is a simple pygame project where i decided to learn pygame and as example, create vampire survivors which is a simple yet interesting game. For this game i decided to use nothing but pygame library. Currently there is only movement, background image and collision detection. Next element is creating simple enemies.